Add MySql connector wrapper, implement appConfig
MySql connector uses the mysql node module to make specified
connections and queries of the application's databases.

AppConfig was finally implemented to read in a config.json file
from the home of the user (`/User/<user>/.` for Mac, `C:/Users/<user>/.`
for Windows, and `/home/<user>/.` for Linux).

The config.json file must be the specified format:

```
{
  "adminEmails": [
    "test@test.com"
  ],
  "db": {
    "host": "the database host",
    "user": "the database user",
    "password": "the database password",
    "database": "the database itself"
  }
}
```